[RENAME] serialize.py to serializers.py
